+++ src/firebase/index.ts
@@ -5,5 +5,5 @@ export {
   type InitFirebaseOptions,
 } from './client';
 export { caspianCollections, type CaspianCollections } from './collections';
-export { CASPIAN_FIRESTORE_RULES } from './rules';
+export { CASPIAN_FIRESTORE_RULES, CASPIAN_STORAGE_RULES } from './rules';
 export { CASPIAN_FIRESTORE_INDEXES } from './indexes';